4.21*: bloglist expansion, step9: Add check a that a user who added the blog is allowed to delete it

+++ part4/blog/controllers/blogs.js
@@ -10,7 +10,9 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const token = request.token
+  console.log(token)
   const decodedToken = jwt.verify(token, process.env.SECRET)
+  console.log('token', decodedToken)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
@@ -39,6 +41,16 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const blog = await Blog.findById(request.params.id).populate('user')
+
+  if (blog.user._id.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'unauthorised operation' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
